Change datatype of Content.length to long long (#397)
* Change datatype of Content.length to long long

---------

+++ blacksheep/contents.pyx
@@ -31,7 +31,7 @@ cdef class StreamedContent(Content):
         self,
         bytes content_type,
         object data_provider,
-        int data_length = -1
+        long long data_length = -1
     ):
         self.type = content_type
         self.body = None
